add Logic Generation to Generator tool

+++ Intotech.GhostRider/Intotech.GhostRider/GeneratorRealization.cs
@@ -1,4 +1,5 @@
-﻿using Intotech.ReflectiveTools.SourceGenerators.Models2DtosGenerator;
+﻿using Intotech.ReflectiveTools.SourceGenerators.LogicGenerator;
+using Intotech.ReflectiveTools.SourceGenerators.Models2DtosGenerator;
 using Intotech.ReflectiveTools.SourceGenerators.ModelsToDtoGenerator;
 using System;
 using System.Collections.Generic;
@@ -62,18 +63,88 @@ namespace Intotech.GhostRider
                 return true;
             }
         }
-        public static bool FolderCleaner(string outputDirectory)
+        public bool LogicRender(string inputDllPath, string outputDirectory, string usings, string nameSpace)
+        {
+            List<string> selectedObj = new() { "Intotech.Wheelo.Bll.Persistence.csproj", "Logic.cs" };
+
+            if (Directory.GetFiles(outputDirectory).Length - selectedObj.Count() == 0)
+            {
+                try
+                {
+                    LogicRendererRunner.LoadAndReadAssembly(inputDllPath, outputDirectory, usings, nameSpace, false);
+                }
+                catch (Exception ex)
+                {
+                    MessageBox.Show(ex.ToString());
+                    throw;
+                }
+                return false;
+            }
+            else
+            {
+                
+
+                if (FolderCleaner(outputDirectory, selectedObj) == false)
+                {
+                    return false;
+                }
+
+
+                return true;
+            }
+        }
+        public bool ILogicRender(string inputDllPath, string outputDirectory, string usings, string nameSpace)
+        {
+            if (Directory.GetFiles(outputDirectory).Length == 0)
+            {
+                try
+                {
+                    LogicRendererRunner.LoadAndReadAssembly(inputDllPath, outputDirectory, usings, nameSpace, true);
+                }
+                catch (Exception ex)
+                {
+                    MessageBox.Show(ex.ToString());
+                    throw;
+                }
+                return false;
+            }
+            else
+            {
+                if (FolderCleaner(outputDirectory) == false)
+                {
+                    return false;
+                }
+
+
+                return true;
+            }
+        }
+
+        public static bool FolderCleaner(string outputDirectory, List<string> selectedObj = null)
         {
 
             DialogResult result = MessageBox.Show("There are objects in the folder, delete them?", "", MessageBoxButtons.YesNo);
 
             if (result == DialogResult.Yes)
             {
+                
                 DirectoryInfo directory = new DirectoryInfo(outputDirectory);
 
                 foreach (FileInfo file in directory.GetFiles())
                 {
-                    file.Delete();
+                    if (selectedObj != null)
+                    {
+                        if (!selectedObj.Contains(file.Name))
+                        {
+                            file.Delete();
+                        }
+                    }
+                    else
+                    {
+                        file.Delete();
+                    }
+                    
+                    
                 }
 
                 Console.WriteLine($"Files in path {outputDirectory}, deleted.");
